move getBoundRegisteredWorkers
